requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
fonttools==4.28.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.1.1
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
setuptools-scm==6.3.2
six==1.16.0
SQLAlchemy==1.4.27
tomli==2.0.1
urllib3==1.26.7
Werkzeug==2.0.2
psycopg2-binary==2.9.3
